+++ src/kelary/IviVisaNetSample/GacLoader.cs
@@ -1,6 +1,10 @@
 using System;
 using System.Linq;
+using System.Net;
 using Ivi.Visa.ConflictManager;
+
+#pragma warning disable CA1305
+#pragma warning disable IDE0150
 
 #nullable enable
 
@@ -127,5 +131,161 @@ public static class GacLoader
                 Console.WriteLine( $"Failed to load assembly {visaLibrary.FriendlyName}: {exception.Message}" );
             }
         }
+    }
+
+    /// <summary>   Queries the instrument identity string. </summary>
+    /// <remarks>   2025-08-12. </remarks>
+    /// <exception cref="ArgumentException">            Thrown when one or more arguments have
+    ///                                                 unsupported or illegal values. </exception>
+    /// <exception cref="InvalidOperationException">    Thrown when the requested operation is
+    ///                                                 invalid. </exception>
+    /// <param name="resourceName"> Name of the resource. </param>
+    /// <param name="verbose">      (Optional) True to verbose. </param>
+    /// <returns>   The identity. </returns>
+    public static string QueryIdentity( string resourceName, bool verbose = false )
+    {
+        if ( string.IsNullOrWhiteSpace( resourceName ) )
+            throw new ArgumentException( $"{nameof( resourceName )} cannot be null or empty.", nameof( resourceName ) );
+        // Connect to the instrument.
+        if ( verbose ) Console.WriteLine( $"Opening a VISA session to '{resourceName}'..." );
+        using Ivi.Visa.IVisaSession visaSession = Ivi.Visa.GlobalResourceManager.Open( resourceName, Ivi.Visa.AccessModes.ExclusiveLock, 2000 );
+        if ( visaSession is Ivi.Visa.IMessageBasedSession messageBasedSession )
+        {
+            // Ensure termination character is enabled as here in example we use a SOCKET connection.
+            messageBasedSession.TerminationCharacterEnabled = true;
+            // Request information about an instrument.
+            if ( verbose ) Console.WriteLine( "\tReading instrument identification string..." );
+            messageBasedSession.FormattedIO.WriteLine( "*IDN?" );
+            return messageBasedSession.FormattedIO.ReadLine();
+        }
+        else
+        {
+            throw new InvalidOperationException( "Not a message-based session." );
+        }
+    }
+
+    /// <summary>   Identify visa session. </summary>
+    /// <remarks>   2025-08-13. </remarks>
+    /// <param name="resourceName"> Name of the resource. </param>
+    /// <returns>   A string. </returns>
+    public static string IdentifyVisaSession( string resourceName )
+    {
+        if ( string.IsNullOrWhiteSpace( resourceName ) )
+        {
+            return $"{nameof( resourceName )} is null or empty or white space.";
+        }
+
+        if ( resourceName.StartsWith( "TCPIP", StringComparison.OrdinalIgnoreCase ) )
+        {
+            System.Text.StringBuilder sb = new();
+            _ = sb.AppendLine( $"Visa Session for '{resourceName}': " );
+
+            using Ivi.Visa.IVisaSession visaSession = Ivi.Visa.GlobalResourceManager.Open( resourceName, Ivi.Visa.AccessModes.ExclusiveLock, 2000 );
+
+            // MessageBasedSession, ITcpipSession2, ITcpipSession, IMessageBasedSession, IVisaSession, IDisposable
+
+            if ( visaSession is Ivi.Visa.IMessageBasedSession )
+            {
+                _ = sb.AppendLine( $"\tis a '{nameof( Ivi )}.{nameof( Ivi.Visa )}.{nameof( Ivi.Visa.IMessageBasedSession )}'." );
+
+                _ = sb.Append( visaSession is Keysight.Visa.MessageBasedSession
+                    ? $"\tis"
+                    : $"\tis not" );
+                _ = sb.AppendLine( $" a '{nameof( Keysight )}.{nameof( Keysight.Visa )}.{nameof( Keysight.Visa.MessageBasedSession )}'." );
+
+                _ = sb.Append( visaSession is Keysight.Visa.TcpipSession
+                    ? $"\tis"
+                    : $"\tis not" );
+                _ = sb.AppendLine( $" a '{nameof( Keysight )}.{nameof( Keysight.Visa )}.{nameof( Keysight.Visa.TcpipSession )}'." );
+
+                _ = sb.Append( visaSession is Ivi.Visa.IVisaSession
+                    ? $"\tis"
+                    : $"\tis not" );
+                _ = sb.AppendLine( $" a '{nameof( Ivi )}.{nameof( Ivi.Visa )}.{nameof( Ivi.Visa.IVisaSession )}'." );
+
+                _ = sb.Append( visaSession is Ivi.Visa.ITcpipSession
+                    ? $"\tis"
+                    : $"\tis not" );
+                _ = sb.AppendLine( $" a '{nameof( Ivi )}.{nameof( Ivi.Visa )}.{nameof( Ivi.Visa.ITcpipSession )}'." );
+
+#if NET9_0_OR_GREATER
+                string typeName = "Ivi.Visa.ITcpipSession2";
+                Type? iTcpIpSession2 = Type.GetType( typeName, false, true );
+                if ( iTcpIpSession2 == null )
+                    _ = sb.AppendLine( $"\tThe type '{typeName}' does not exist in the namespace '{nameof( Ivi )}.{nameof( Ivi.Visa )}'." );
+                else
+                {
+                    try
+                    {
+                        _ = sb.Append( iTcpIpSession2 is not null
+                            ? $"\tis"
+                            : $"\tis not" );
+                        _ = sb.AppendLine( $" a 'typeName'." );
+                    }
+                    catch ( Exception ex )
+                    {
+                        _ = sb.AppendLine( $"\tException handling 'typeName':\n\t\t{ex.Message}." );
+                    }
+                }
+
+#else
+                try
+                {
+                    _ = sb.Append( visaSession is Ivi.Visa.ITcpipSession2
+                        ? $"\tis"
+                        : $"\tis not" );
+                    _ = sb.AppendLine( $" a '{nameof( Ivi )}.{nameof( Ivi.Visa )}.{nameof( Ivi.Visa.ITcpipSession2 )}'." );
+                }
+                catch ( Exception ex )
+                {
+                    _ = sb.AppendLine( $" Exception handling '{nameof( Ivi )}.{nameof( Ivi.Visa )}.{nameof( Ivi.Visa.ITcpipSession2 )}'; {ex.Message}." );
+                }
+#endif
+            }
+            else
+            {
+                _ = sb.AppendLine( $"\tis not a '{nameof( Ivi )}.{nameof( Ivi.Visa )}.{nameof( Ivi.Visa.IMessageBasedSession )}'." );
+            }
+            return sb.ToString();
+        }
+        else
+        {
+            return $"Resource '{resourceName}' is not a TCPIP resource.";
+        }
+
+    }
+
+    /// <summary>   Attempts to ping. </summary>
+    /// <remarks>   2025-08-12. </remarks>
+    /// <param name="resourceName"> Name of the resource. </param>
+    /// <returns>   True if it succeeds, false if it fails. </returns>
+    public static bool TryPing( string resourceName, out string details )
+    {
+        bool outcome = false;
+        try
+        {
+            System.Net.NetworkInformation.Ping ping = new();
+            System.Net.NetworkInformation.PingOptions pingOptions = new( 4, true );
+            byte[] buffer = [0, 0];
+            if ( resourceName.StartsWith( "TCPIP", StringComparison.OrdinalIgnoreCase ) )
+                resourceName = resourceName.Split( ':' )[2];
+            if ( IPAddress.TryParse( resourceName, out IPAddress? address ) )
+            {
+                outcome = ping.Send( address, 1000, buffer, pingOptions ).Status == System.Net.NetworkInformation.IPStatus.Success;
+                details = outcome
+                    ? $"Instrument found at '{resourceName}'."
+                    : $"Attempt Ping instrument at '{resourceName}' failed.";
+            }
+            else
+            {
+                details = $"Non TCPIP Instrument at '{resourceName}'.";
+                outcome = true;
+            }
+        }
+        catch ( Exception )
+        {
+            details = $"Exception occurred pinging {resourceName}; .";
+        }
+        return outcome;
     }
 }